feat: add PublisherModule for RabbitMQ publishing and video metadata events docs: add comprehensive process documentation for Video Platform Server fix: resolve critical issues in RabbitMQ message structure and authentication guards

+++ api_service/src/video/dto/user-videos.response.ts
@@ -17,13 +17,13 @@ export class UserVideoResponse {
   @Field(() => String, { nullable: true })
   thumbnailUrl?: string;
 
-  @Field(() => BigInt)
+  @Field(() => Int)
   videoView!: number;
 
-  @Field(() => BigInt)
+  @Field(() => Int)
   videoLike!: number;
 
-  @Field(() => BigInt)
+  @Field(() => Int)
   videoDislike!: number;
 
   @Field(() => String)
